+++ apps/vs-code-designer/src/app/commands/createNewCodeProject/createCodeProjectSteps/createLogicApp/ScriptProjectCreateStep.ts
@@ -4,6 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 import {
   azureWebJobsStorageKey,
+  funcIgnoreFileName,
   gitignoreFileName,
   hostFileName,
   localEmulatorConnectionString,
@@ -21,6 +22,7 @@ import * as fse from 'fs-extra';
 import * as os from 'os';
 import * as path from 'path';
 import type { Progress } from 'vscode';
+import { getGitIgnoreContent } from '../../../../utils/git';
 
 /**
  * This class represents a step that creates the contents of a new script project.
@@ -79,28 +81,11 @@ export class ScriptProjectCreateStep extends ProjectCodeCreateStepBase {
     // Create the .gitignore file
     const gitignorePath: string = path.join(context.projectPath, gitignoreFileName);
     if (await confirmOverwriteFile(context, gitignorePath)) {
-      await fse.writeFile(
-        gitignorePath,
-        this.gitignore.concat(`
-# Azure Functions artifacts
-bin
-obj
-appsettings.json
-local.settings.json
-__blobstorage__
-.debug
-__queuestorage__
-__azurite_db*__.json
-
-# Added folders and file patterns
-workflow-designtime/
-.vscode/
-*.code-workspace`)
-      );
+      await fse.writeFile(gitignorePath, this.gitignore.concat(getGitIgnoreContent()));
     }
 
     // Create the .funcignore file
-    const funcIgnorePath: string = path.join(context.projectPath, '.funcignore');
+    const funcIgnorePath: string = path.join(context.projectPath, funcIgnoreFileName);
     if (await confirmOverwriteFile(context, funcIgnorePath)) {
       await fse.writeFile(funcIgnorePath, this.funcignore.sort().join(os.EOL));
     }
