+++ apps/vs-code-designer/src/app/commands/createNewProject/createProjectSteps/ScriptProjectCreateStep.ts
@@ -6,6 +6,7 @@ import {
   ProjectDirectoryPath,
   appKindSetting,
   azureWebJobsStorageKey,
+  funcIgnoreFileName,
   gitignoreFileName,
   hostFileName,
   localSettingsFileName,
@@ -24,6 +25,7 @@ import * as fse from 'fs-extra';
 import * as os from 'os';
 import * as path from 'path';
 import type { Progress } from 'vscode';
+import { getGitIgnoreContent } from '../../../utils/git';
 
 export class ScriptProjectCreateStep extends ProjectCreateStepBase {
   protected funcignore: string[] = [
@@ -81,27 +83,10 @@ export class ScriptProjectCreateStep extends ProjectCreateStepBase {
     const gitignorePath = path.join(baseDirectory, gitignoreFileName);
 
     if (await confirmOverwriteFile(context, gitignorePath)) {
-      await fse.writeFile(
-        gitignorePath,
-        this.gitignore.concat(`
-# Azure Functions artifacts
-bin
-obj
-appsettings.json
-local.settings.json
-__blobstorage__
-.debug
-__queuestorage__
-__azurite_db*__.json
-
-# Added folders and file patterns
-workflow-designtime/
-.vscode/
-*.code-workspace`)
-      );
+      await fse.writeFile(gitignorePath, this.gitignore.concat(getGitIgnoreContent()));
     }
 
-    const funcIgnorePath: string = path.join(context.projectPath, '.funcignore');
+    const funcIgnorePath: string = path.join(context.projectPath, funcIgnoreFileName);
     if (await confirmOverwriteFile(context, funcIgnorePath)) {
       await fse.writeFile(funcIgnorePath, this.funcignore.sort().join(os.EOL));
     }
