+++ apps/vs-code-designer/src/app/utils/git.ts
@@ -2,9 +2,11 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-import { gitCommand } from '../../constants';
+import { gitCommand, gitignoreFileName } from '../../constants';
 import { ext } from '../../extensionVariables';
 import { executeCommand } from './funcCoreTools/cpUtils';
+import * as path from 'path';
+import * as fse from 'fs-extra';
 
 /**
  * Checks if git is installed.
@@ -41,3 +43,30 @@ export async function isInsideRepo(workingDirectory: string): Promise<boolean> {
     return false;
   }
 }
+
+export const removeFromGitIgnore = async (workspacePath: string, pattern: RegExp) => {
+  const gitignorePath: string = path.join(workspacePath, gitignoreFileName);
+  if (await fse.pathExists(gitignorePath)) {
+    let gitignoreContents: string = (await fse.readFile(gitignorePath)).toString();
+    gitignoreContents = gitignoreContents.replace(pattern, '');
+    await fse.writeFile(gitignorePath, gitignoreContents);
+  }
+};
+
+export const getGitIgnoreContent = () => {
+  return `
+# Azure logic apps artifacts
+bin
+obj
+appsettings.json
+local.settings.json
+__blobstorage__
+.debug
+__queuestorage__
+__azurite_db*__.json
+
+# Added folders and file patterns
+workflow-designtime/
+.vscode/
+*.code-workspace`;
+};
